Add Images and Add Text Initial Changes
Add the partial implementation which allows the user to select both
images and text for the bingo squares. Still need to implement the
add numbers functionality, as well as add the ability to create and
download bingo cards. Will probably want to clean up the code and
pretty up the display as well.